Fix most clippy warnings
