+++ hacspec-scrambledb/hash-to-curve/src/expand_message.rs
@@ -1,132 +1,132 @@
-//!   `expand_message` is a function that generates a uniformly random byte
-//!    string.  It takes three arguments:
+//! `expand_message` is a function that generates a uniformly random byte
+//! string.  It takes three arguments:
 //!
-//!    1.  `msg`, a byte string containing the message to hash,
+//! 1.  `msg`, a byte string containing the message to hash,
 //!
-//!    2.  `DST`, a byte string that acts as a domain separation tag, and
+//! 2.  `DST`, a byte string that acts as a domain separation tag, and
 //!
-//!    3.  `len_in_bytes`, the number of bytes to be generated.
+//! 3.  `len_in_bytes`, the number of bytes to be generated.
 //!
-//!    This document defines the following two variants of expand_message:
+//! This document defines the following two variants of expand_message:
 //!
-//!    *  [expand_message_xmd] is appropriate for use with a
-//!       wide range of hash functions, including SHA-2 [FIPS180-4], SHA-3
-//!       [FIPS202], BLAKE2 [RFC7693], and others.
+//! *  [expand_message_xmd] is appropriate for use with a
+//!    wide range of hash functions, including SHA-2 [FIPS180-4], SHA-3
+//!    [FIPS202], BLAKE2 [RFC7693], and others.
 //!
-//!    *  [expand_message_xof]  appropriate for use with
-//!       extendable-output functions (XOFs) including functions in the
-//!       SHAKE [FIPS202] or BLAKE2X [BLAKE2X] families.
+//! *  [expand_message_xof]  appropriate for use with
+//!    extendable-output functions (XOFs) including functions in the
+//!    SHAKE [FIPS202] or BLAKE2X [BLAKE2X] families.
 //!
-//!    These variants should suffice for the vast majority of use cases, but
-//!    other variants are possible.
+//! These variants should suffice for the vast majority of use cases, but
+//! other variants are possible.
 //!
 //! ### Defining other expand_message variants
 //!
-//!    When defining a new `expand_message` variant, the most important
-//!    consideration is that `hash_to_field` models `expand_message` as a random
-//!    oracle.  Thus, implementors SHOULD prove indifferentiability from a
-//!    random oracle under an appropriate assumption about the underlying
-//!    cryptographic primitives; see Section 10.5 for more information.
+//! When defining a new `expand_message` variant, the most important
+//! consideration is that `hash_to_field` models `expand_message` as a random
+//! oracle.  Thus, implementors SHOULD prove indifferentiability from a
+//! random oracle under an appropriate assumption about the underlying
+//! cryptographic primitives; see Section 10.5 for more information.
 //!
-//!    In addition, `expand_message` variants:
+//! In addition, `expand_message` variants:
 //!
-//!    *  MUST give collision resistance commensurate with the security
-//!       level of the target elliptic curve.
+//! *  MUST give collision resistance commensurate with the security
+//!    level of the target elliptic curve.
 //!
-//!    *  MUST be built on primitives designed for use in applications
-//!       requiring cryptographic randomness.  As examples, a secure stream
-//!       cipher is an appropriate primitive, whereas a Mersenne twister
-//!       pseudorandom number generator [MT98] is not.
+//! *  MUST be built on primitives designed for use in applications
+//!    requiring cryptographic randomness.  As examples, a secure stream
+//!    cipher is an appropriate primitive, whereas a Mersenne twister
+//!    pseudorandom number generator [MT98] is not.
 //!
-//!    *  MUST NOT use rejection sampling.
+//! *  MUST NOT use rejection sampling.
 //!
-//!    *  MUST give independent values for distinct (`msg`, `DST`, `length`)
-//!       inputs.  Meeting this requirement is subtle.  As a simplified
-//!       example, hashing `msg || DST` does not work, because in this case
-//!       distinct `(msg, DST)` pairs whose concatenations are equal will
-//!       return the same output (e.g., `("AB", "CDEF")` and `("ABC", "DEF")`).
-//!       The variants defined in this document use a suffix-free encoding
-//!       of `DST` to avoid this issue.
+//! *  MUST give independent values for distinct (`msg`, `DST`, `length`)
+//!    inputs.  Meeting this requirement is subtle.  As a simplified
+//!    example, hashing `msg || DST` does not work, because in this case
+//!    distinct `(msg, DST)` pairs whose concatenations are equal will
+//!    return the same output (e.g., `("AB", "CDEF")` and `("ABC", "DEF")`).
+//!    The variants defined in this document use a suffix-free encoding
+//!    of `DST` to avoid this issue.
 //!
-//!    *  MUST use the domain separation tag `DST` to ensure that invocations
-//!       of cryptographic primitives inside of `expand_message` are domain
-//!       separated from invocations outside of `expand_message`.  For
-//!       example, if the `expand_message` variant uses a hash function H, an
-//!       encoding of `DST` MUST be added either as a prefix or a suffix of
-//!       the input to each invocation of `H`.  Adding `DST` as a suffix is the
-//!       RECOMMENDED approach.
+//! *  MUST use the domain separation tag `DST` to ensure that invocations
+//!    of cryptographic primitives inside of `expand_message` are domain
+//!    separated from invocations outside of `expand_message`.  For
+//!    example, if the `expand_message` variant uses a hash function H, an
+//!    encoding of `DST` MUST be added either as a prefix or a suffix of
+//!    the input to each invocation of `H`.  Adding `DST` as a suffix is the
+//!    RECOMMENDED approach.
 //!
-//!    *  SHOULD read `msg` exactly once, for efficiency when `msg` is long.
+//! *  SHOULD read `msg` exactly once, for efficiency when `msg` is long.
 //!
-//!    In addition, each `expand_message` variant MUST specify a unique
-//!    `EXP_TAG` that identifies that variant in a Suite ID.  See Section 8.10
-//!    for more
+//! In addition, each `expand_message` variant MUST specify a unique
+//! `EXP_TAG` that identifies that variant in a Suite ID.  See Section 8.10
+//! for more
 //!
 //! ### Using DSTs longer than 255 bytes
 //!
-//!    The `expand_message` variants defined in this section accept domain
-//!    separation tags of at most 255 bytes.  If applications require a
-//!    domain separation tag longer than 255 bytes, e.g., because of
-//!    requirements imposed by an invoking protocol, implementors MUST
-//!    compute a short domain separation tag by hashing, as follows:
+//! The `expand_message` variants defined in this section accept domain
+//! separation tags of at most 255 bytes.  If applications require a
+//! domain separation tag longer than 255 bytes, e.g., because of
+//! requirements imposed by an invoking protocol, implementors MUST
+//! compute a short domain separation tag by hashing, as follows:
 //!
-//!    *  For expand_message_xmd using hash function H, DST is computed as
+//! *  For expand_message_xmd using hash function H, DST is computed as
 //!
 //!
-//! 	  ``` text
-//! 	  DST = H("H2C-OVERSIZE-DST-" || a_very_long_DST)
-//! 	  ```
+//! ``` text
+//! DST = H("H2C-OVERSIZE-DST-" || a_very_long_DST)
+//! ```
 //!
-//!    *  For expand_message_xof using extendable-output function H, DST is
-//!       computed as
+//! *  For expand_message_xof using extendable-output function H, DST is
+//!    computed as
 //!
-//! 	  ``` text
-//!       DST = H("H2C-OVERSIZE-DST-" || a_very_long_DST, ceil(2 * k / 8))
-//! 	  ```
+//! ``` text
+//! DST = H("H2C-OVERSIZE-DST-" || a_very_long_DST, ceil(2 * k / 8))
+//! ```
 //!
-//!    Here, a_very_long_DST is the DST whose length is greater than 255
-//!    bytes, "H2C-OVERSIZE-DST-" is a 17-byte ASCII string literal, and k
-//!    is the target security level in bits.
+//! Here, a_very_long_DST is the DST whose length is greater than 255
+//! bytes, "H2C-OVERSIZE-DST-" is a 17-byte ASCII string literal, and k
+//! is the target security level in bits.
 
 use crate::{hacspec_helper::FunctionalVec, hasher::HashAlgorithm, Error};
 
-///    The `expand_message_xof` function produces a uniformly random byte
-///    string using an extendable-output function (XOF) H.  For security, H
-///    MUST meet the following criteria:
+/// The `expand_message_xof` function produces a uniformly random byte
+/// string using an extendable-output function (XOF) H.  For security, H
+/// MUST meet the following criteria:
 ///
-///    *  The collision resistance of H MUST be at least k bits.
+/// *  The collision resistance of H MUST be at least k bits.
 ///
-///    *  H MUST be an XOF that has been proved indifferentiable from a
-///       random oracle under a reasonable cryptographic assumption.
+/// *  H MUST be an XOF that has been proved indifferentiable from a
+///    random oracle under a reasonable cryptographic assumption.
 ///
-///    The SHAKE [FIPS202] XOF family is a typical and RECOMMENDED choice.
-///    As an example, for 128-bit security, SHAKE128 would be an appropriate
-///    choice.
+/// The SHAKE [FIPS202] XOF family is a typical and RECOMMENDED choice.
+/// As an example, for 128-bit security, SHAKE128 would be an appropriate
+/// choice.
 ///
-///    The following procedure implements `expand_message_xof`.
+/// The following procedure implements `expand_message_xof`.
 ///
 /// ```text
-///    expand_message_xof(msg, DST, len_in_bytes)
+/// expand_message_xof(msg, DST, len_in_bytes)
 ///
-///    Parameters:
-///    - H(m, d), an extendable-output function that processes
-///               input message m and returns d bytes.
+/// Parameters:
+/// - H(m, d), an extendable-output function that processes
+///   input message m and returns d bytes.
 ///
-///    Input:
-///    - msg, a byte string.
-///    - DST, a byte string of at most 255 bytes.
-///      See below for information on using longer DSTs.
-///    - len_in_bytes, the length of the requested output in bytes.
+/// Input:
+/// - msg, a byte string.
+/// - DST, a byte string of at most 255 bytes.
+///   See below for information on using longer DSTs.
+/// - len_in_bytes, the length of the requested output in bytes.
 ///
-///    Output:
-///    - uniform_bytes, a byte string.
+/// Output:
+/// - uniform_bytes, a byte string.
 ///
-///    Steps:
-///    1. ABORT if len_in_bytes > 65535 or len(DST) > 255
-///    2. DST_prime = DST || I2OSP(len(DST), 1)
-///    3. msg_prime = msg || I2OSP(len_in_bytes, 2) || DST_prime
-///    4. uniform_bytes = H(msg_prime, len_in_bytes)
-///    5. return uniform_bytes
+/// Steps:
+/// 1. ABORT if len_in_bytes > 65535 or len(DST) > 255
+/// 2. DST_prime = DST || I2OSP(len(DST), 1)
+/// 3. msg_prime = msg || I2OSP(len_in_bytes, 2) || DST_prime
+/// 4. uniform_bytes = H(msg_prime, len_in_bytes)
+/// 5. return uniform_bytes
 /// ```
 #[allow(non_snake_case, unused)]
 pub fn expand_message_xof(
@@ -140,66 +140,66 @@ pub fn expand_message_xof(
     todo!()
 }
 
+
+/// The `expand_message_xmd` function produces a uniformly random byte
+/// string using a cryptographic hash function H that outputs b bits.
+/// For security, H MUST meet the following requirements:
 ///
-///    The `expand_message_xmd` function produces a uniformly random byte
-///    string using a cryptographic hash function H that outputs b bits.
-///    For security, H MUST meet the following requirements:
+/// *  The number of bits output by H MUST be b >= 2 * k, for k the
+///    target security level in bits, and b MUST be divisible by 8.  The
+///    first requirement ensures k-bit collision resistance; the second
+///    ensures uniformity of `expand_message_xmd`'s output.
 ///
-///    *  The number of bits output by H MUST be b >= 2 * k, for k the
-///       target security level in bits, and b MUST be divisible by 8.  The
-///       first requirement ensures k-bit collision resistance; the second
-///       ensures uniformity of `expand_message_xmd`'s output.
+/// *  H MAY be a Merkle-Damgaard hash function like SHA-2.  In this
+///    case, security holds when the underlying compression function is
+///    modeled as a random oracle [CDMP05].  (See Section 10.6 for
+///    discussion.)
 ///
-///    *  H MAY be a Merkle-Damgaard hash function like SHA-2.  In this
-///       case, security holds when the underlying compression function is
-///       modeled as a random oracle [CDMP05].  (See Section 10.6 for
-///       discussion.)
+/// *  H MAY be a sponge-based hash function like SHA-3 or BLAKE2.  In
+///    this case, security holds when the inner function is modeled as a
+///    random transformation or as a random permutation [BDPV08].
 ///
-///    *  H MAY be a sponge-based hash function like SHA-3 or BLAKE2.  In
-///       this case, security holds when the inner function is modeled as a
-///       random transformation or as a random permutation [BDPV08].
+/// *  Otherwise, H MUST be a hash function that has been proved
+///    indifferentiable from a random oracle [MRH04] under a reasonable
+///    cryptographic assumption.
 ///
-///    *  Otherwise, H MUST be a hash function that has been proved
-///       indifferentiable from a random oracle [MRH04] under a reasonable
-///       cryptographic assumption.
+/// SHA-2 [FIPS180-4] and SHA-3 [FIPS202] are typical and RECOMMENDED
+/// choices.  As an example, for the 128-bit security level, b >= 256
+/// bits and either SHA-256 or SHA3-256 would be an appropriate choice.
 ///
-///    SHA-2 [FIPS180-4] and SHA-3 [FIPS202] are typical and RECOMMENDED
-///    choices.  As an example, for the 128-bit security level, b >= 256
-///    bits and either SHA-256 or SHA3-256 would be an appropriate choice.
+/// The hash function H is assumed to work by repeatedly ingesting fixed-
+/// length blocks of data.  The length in bits of these blocks is called
+/// the input block size (s).  As examples, s = 1024 for SHA-512
+/// [FIPS180-4] and s = 576 for SHA3-512 [FIPS202].  For correctness, H
+/// requires b <= s.
 ///
-///    The hash function H is assumed to work by repeatedly ingesting fixed-
-///    length blocks of data.  The length in bits of these blocks is called
-///    the input block size (s).  As examples, s = 1024 for SHA-512
-///    [FIPS180-4] and s = 576 for SHA3-512 [FIPS202].  For correctness, H
-///    requires b <= s.
-///
-///    The following procedure implements `expand_message_xmd`.
+/// The following procedure implements `expand_message_xmd`.
 /// ```text
-///    expand_message_xmd(msg, DST, len_in_bytes)
+/// expand_message_xmd(msg, DST, len_in_bytes)
 ///
-///    Parameters:
-///    - H, a hash function (see requirements above).
-///    - b_in_bytes, b / 8 for b the output size of H in bits.
-///      For example, for b = 256, b_in_bytes = 32.
-///    - s_in_bytes, the input block size of H, measured in bytes (see
-///      discussion above). For example, for SHA-256, s_in_bytes = 64.
+/// Parameters:
+/// - H, a hash function (see requirements above).
+/// - b_in_bytes, b / 8 for b the output size of H in bits.
+///   For example, for b = 256, b_in_bytes = 32.
+/// - s_in_bytes, the input block size of H, measured in bytes (see
+///   discussion above). For example, for SHA-256, s_in_bytes = 64.
 ///
-///    Input:
-///    - msg, a byte string.
-///    - DST, a byte string of at most 255 bytes.
-///      See below for information on using longer DSTs.
-///    - len_in_bytes, the length of the requested output in bytes,
-///      not greater than the lesser of (255 * b_in_bytes) or 2^16-1.
+/// Input:
+/// - msg, a byte string.
+/// - DST, a byte string of at most 255 bytes.
+///   See below for information on using longer DSTs.
+/// - len_in_bytes, the length of the requested output in bytes,
+///   not greater than the lesser of (255 * b_in_bytes) or 2^16-1.
 ///
-///    Output:
-///    - uniform_bytes, a byte string.
+/// Output:
+/// - uniform_bytes, a byte string.
 ///
-///    Steps:
-///    1.  ell = ceil(len_in_bytes / b_in_bytes)
-///    2.  ABORT if ell > 255 or len_in_bytes > 65535 or len(DST) > 255
-///    3.  DST_prime = DST || I2OSP(len(DST), 1)
-///    4.  Z_pad = I2OSP(0, s_in_bytes)
-///    5.  l_i_b_str = I2OSP(len_in_bytes, 2)
+/// Steps:
+/// 1.  ell = ceil(len_in_bytes / b_in_bytes)
+/// 2.  ABORT if ell > 255 or len_in_bytes > 65535 or len(DST) > 255
+/// 3.  DST_prime = DST || I2OSP(len(DST), 1)
+/// 4.  Z_pad = I2OSP(0, s_in_bytes)
+/// 5.  l_i_b_str = I2OSP(len_in_bytes, 2)
 ///    6.  msg_prime = Z_pad || msg || l_i_b_str || I2OSP(0, 1) || DST_prime
 ///    7.  b_0 = H(msg_prime)
 ///    8.  b_1 = H(b_0 || I2OSP(1, 1) || DST_prime)
@@ -208,21 +208,21 @@ pub fn expand_message_xof(
 ///    11. uniform_bytes = b_1 || ... || b_ell
 ///    12. return substr(uniform_bytes, 0, len_in_bytes)
 /// ```
-///    Note that the string Z_pad (step 6) is prefixed to msg before
-///    computing b_0 (step 7).  This is necessary for security when H is a
-///    Merkle-Damgaard hash, e.g., SHA-2 (see Section 10.6).  Hashing this
-///    additional data means that the cost of computing b_0 is higher than
-///    the cost of simply computing H(msg).  In most settings this overhead
-///    is negligible, because the cost of evaluating H is much less than the
-///    other costs involved in hashing to a curve.
+/// Note that the string Z_pad (step 6) is prefixed to msg before
+/// computing b_0 (step 7).  This is necessary for security when H is a
+/// Merkle-Damgaard hash, e.g., SHA-2 (see Section 10.6).  Hashing this
+/// additional data means that the cost of computing b_0 is higher than
+/// the cost of simply computing H(msg).  In most settings this overhead
+/// is negligible, because the cost of evaluating H is much less than the
+/// other costs involved in hashing to a curve.
 ///
-///    It is possible, however, to entirely avoid this overhead by taking
-///    advantage of the fact that Z_pad depends only on H, and not on the
-///    arguments to `expand_message_xmd`.  To do so, first precompute and save
-///    the internal state of H after ingesting Z_pad.  Then, when computing
-///    b_0, initialize H using the saved state.  Further details are
-///    implementation dependent, and beyond the scope of this document.
-///
+/// It is possible, however, to entirely avoid this overhead by taking
+/// advantage of the fact that Z_pad depends only on H, and not on the
+/// arguments to `expand_message_xmd`.  To do so, first precompute and save
+/// the internal state of H after ingesting Z_pad.  Then, when computing
+/// b_0, initialize H using the saved state.  Further details are
+/// implementation dependent, and beyond the scope of this document.
+
 #[allow(non_snake_case)]
 pub fn expand_message_xmd<H: HashAlgorithm>(
     msg: &[u8],
