+++ hacspec-scrambledb/oprf/src/util.rs
@@ -1,5 +1,4 @@
-use p256::{NatMod, P256Scalar};
-use rand::prelude::*;
+use p256::P256Scalar;
 
 /// ### 4.7.2 Random Number Generation Using Extra Random Bits
 ///
@@ -13,31 +12,5 @@ use rand::prelude::*;
 /// ceil(log2(G.Order())) = 256
 /// ceil(((3 * 256) / 2) / 8) = 48
 pub fn random_scalar() -> P256Scalar {
-    let mut bytes = [0u8; 48];
-    rand::thread_rng().fill_bytes(&mut bytes);
-
-    P256Scalar::from_be_bytes(&bytes)
-}
-
-/// From [RFC8017]:
-///
-/// I2OSP converts a nonnegative integer to an octet string of a
-/// specified length.
-///
-/// ```text
-///    I2OSP (x, xLen)
-///
-///    Input:
-///
-///       x        nonnegative integer to be converted
-///
-///       xLen     intended length of the resulting octet string
-///
-///    Output:
-///
-///          X corresponding octet string of length xLen
-/// ```
-pub fn i2osp(x: usize, x_len: usize) -> Vec<u8> {
-    assert!(x_len <= 8);
-    Vec::from(&x.to_be_bytes()[(8 - x_len)..8])
+    scrambledb_util::random_scalar()
 }
