+++ hacspec-scrambledb/scrambledb-util/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "scrambledb-util"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+p256 = { path = "../p256" }
+rand = "0.8.5"
